backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
PyJWT
email-validator